requirements.txt
aiosmtpd==1.4.6
annotated-types==0.7.0
asgiref==3.8.1
atpublic==5.1
attrs==25.1.0
babel==2.17.0
Django==5.1.5
django-bootstrap-datepicker-plus==5.0.5
django-colorfield==0.12.0
django-recaptcha==4.0.0
django-tinymce==4.1.0
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
